fix: rename filename protected-question-router -> protected-example-router
