+++ template/addons/trpc/auth-index-router.ts
@@ -3,7 +3,7 @@ import { createRouter } from "./context";
 import superjson from "superjson";
 
 import { exampleRouter } from "./example";
-import { protectedExampleRouter } from "./protected-question-router";
+import { protectedExampleRouter } from "./protected-example-router";
 
 export const appRouter = createRouter()
   .transformer(superjson)
