+++ template/addons/trpc/protected-example-router.ts
